move from FeeBuilder to dedicated xcmPaymentApi directory
