+++ .vscode/settings.json
@@ -76,6 +76,7 @@
     "shiden",
     "sibl",
     "statemint",
+    "sufficients",
     "suri",
     "syncpack",
     "teer",
