+++ packages/builder/src/fee/xcmPaymentApi/index.ts
@@ -1,0 +1,1 @@
+export * from './xcmPaymentApi';
